Cleaning up errorprovider
